feat(movies): move search logic to standalone hook
